requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
Pillow>=10.0.0
python-dotenv>=1.0.0
web3>=6.0.0
eth-account==0.8.0
bcrypt==4.0.1
email-validator==2.0.0.post2
python-slugify==8.0.1
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python==4.8.0.74
moviepy==1.0.3